User roles added

+++ app/controllers/wikis_controller.rb
@@ -1,4 +1,5 @@
 class WikisController < ApplicationController
+
   def index
     @wikis = Wiki.all
   end
@@ -9,52 +10,56 @@ class WikisController < ApplicationController
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
-    @wiki = Wiki.new(wiki_params)
-    @wiki.user = current_user
-
-    if @wiki.save
-      flash[:notice] = "Wiki was saved"
-      redirect_to @wiki
-    else
-      flash[:alert] = "There was an error saving the wiki. Please try again."
-      render :new
-    end
+   @wiki = Wiki.new
+   authorize @wiki
+   @wiki.title = params[:wiki][:title]
+   @wiki.body = params[:wiki][:body]
+   @wiki.private = params[:wiki][:private]
+   @wiki.user = current_user
+   if @wiki.save
+     flash[:notice] = "Your wiki was saved."
+     redirect_to @wiki
+   else
+     flash[:alert] = "Something went wrong. Please try again."
+     render :new
+   end
   end
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
-    @wiki.assign_attributes(wiki_params)
-
+    authorize @wiki
+    @wiki.title = params[:wiki][:title]
+    @wiki.body = params[:wiki][:body]
+    @wiki.private = params[:wiki][:private]
+    @wiki.user = current_user
     if @wiki.save
-      flash[:notice] = "Wiki was updated."
-      redirect_to [@wiki]
+      flash[:notice] = "Your wiki was updated."
+      redirect_to @wiki
     else
-      flash.now[:alert] = "There was an error saving the wiki. Please try again."
+      flash.now[:alert] = "Your change was not saved. Please try again."
       render :edit
     end
   end
 
   def destroy
-      @wiki = Wiki.find(params[:id])
-
-      if @wiki.destroy
-        flash[:notice] = "\"#{@wiki.title}\" was successfully deleted."
-        redirect_to @wiki
-      else
-        flash.now[:alert] = "There was an error deleting the wiki."
-        render :show
-      end
-    end
-
-private
-    def wiki_params
-      params.require(:wiki).permit(:title, :body, :private)
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
+    if @wiki.destroy
+      flash[:notice] = "\"#{@wiki.title}\" was deleted!"
+      redirect_to wikis_path
+    else
+      flash.now[:alert] = "Unable to delete, please try again."
+      render :show
     end
   end
+
+end
